1 - Respawn timer is 3 seconds if no bosses are alive and no events are active. Respawn timer is 6 seconds if no bosses are alive and an event is active. 2 - Edited several projectile spread calculations. 3 - Fixed an SCal dart spread attack. 4 - Gave SCal smoother movement. 5 - Queen Slime AI adjustments. 6 - Skeletron Prime movement adjustments. 7 - Ceaseless Void movement adjustments. 8 - Fairies now properly fly away if you take off the Fairy Boots while they're flying near you.

+++ Items/Weapons/Melee/GaelsGreatsword.cs
@@ -132,7 +132,7 @@ namespace CalamityMod.Items.Weapons.Melee
                     float rotation = MathHelper.ToRadians(10f);
                     for (int i = 0; i < numProj; i++)
                     {
-                        Vector2 perturbedSpeed = velocity.RotatedBy(MathHelper.Lerp(-rotation, rotation, i / (numProj - 1)));
+                        Vector2 perturbedSpeed = velocity.RotatedBy(MathHelper.Lerp(-rotation, rotation, i / (float)(numProj - 1)));
                         Projectile.NewProjectile(source, position, perturbedSpeed, type, damage, knockback, player.whoAmI);
                     }
                     break;
